Adds new method for adding components and updates constructor and setter.
New method for adding components to the componentsInSystem. This also updates the component's system variable so that the data between the two objects is in sync.

Updated the setSystemName method to use the constant defined in the Game object for name validation.

Changed datatype for componentsInSystem to ArrayList as there will not be a fixed number of components contained within.