Updates the constructor, setters, and removes redundant code.
Constructor signature and body updated to only require the square name. The position of the square will be set by the Board object, therefore squarePosition was removed from the constructor.

The validation within the setters has been updated to validate against constants set in the Game object.

The variable squareSystem and associated methods have been removed. This is because a square cannot be part of a System, only Components can do this.