Board object has been re-written to enable it to dynamically create Squares and Systems. Tests written to ensure class performs as expected.
The Board object contains two fixed arrays, set by constants in the Game object. These are, squares, and, systems.

The createSquare method is overloaded, the first creates a basic Square object. The second, creates a Component object. Both are treated as Squares within the Board object. To access component specific functionality the squares will need to be cast as components.

Square position is dynamically assigned dependent upon when a square was created (i.e. it is the element position). Error handling has been built in to prevent a board from creating more squares than is permitted.

createSystem is a single method and is respond for creating new ArtemisSystem objects and storing them in the Board object. Constraints have been setup which are determined by the Game object.