Implements tradeComponent method and relevant testing of trade functionality has been incorporated
In order to facilitate this functionality, a number of helper methods have been added.

updateResources is used to add/subtract resources from a player. This differs from setResources in that it does not override current resources, instead it takes the current balance and adds/subtracts from it.

transferResources is used to transfers resources from one player to another player. An announcement of a transfer is made to the game each time a transfer takes place.

consumeActionPoint is used to update a player's action points after performing an action. This takes the current action point balance and subtracts the number specified.