Implements the trade menu functionality and associated testing
In order to display a list of components that an active player can trade for, a number of helper methods have been implemented. The methods were developed to perform a single function which enables them to be tested easier.

getComponentsForTrading returns a map of key-value (reference-component) pairs containing all possible components the current player can trade for. The components added to this are subject to some constraints: 1) the component is owned by a player; 2) the component is not owned by the current player; and, 3) the current player has sufficient resources to purchase the component.

outputComponentsForTrading takes the above map and outputs a user friendly menu, enabling the user to easily select they component they wish to trade for.

getPlayerComponentSelection is general purpose and accepts a map of components. It prompts the user to select one of the valid options, based on the keys specified by the components map. If an invalid selection is made, the user is prompted again, or requested to type 'end' to return to the main menu.

displayTradeMenu brings together all the above functionality and invokes it in the correct manner to facilitate the trade. If the user inputs a valid component, the trade method within the player object is called. This will prompt the counterparty player to accept or reject the trade.