Adds method to facilitate other players being offered a component when the player that lands on it rejects it.
getHighestRoll loops through the players in the game (with the required resources) asking each if they wish to purchase - a new private class-wide player list variable was created, with a getter and setter to facilitate this. All players that responded yes, with have a virtual dice rolled and the highest will become the owner of the component. If there are joint high rolls, this will loop until there is only one high roller. The method returns the winner of the roll.