Implements the offerComponentToOtherPlayers method.
This is called when the currentPlayer does not wish to purchase the component they landed on. The method will get all other players in the game, ask them if they wish to purchase, roll virtual dice, and the winning player will have the purchaseComponent method invoked. This will give them ownership of the component and update the relevant data points (e.g. player resources, component owner, etc.).

If noone wants to purchase, then the component will remain unowned.