Replaced boolean variable with method in Artemis System
Implemented method to test if system is fully developed i.e. it checks
if all components contained therein are fully developed and returns
either true or false.

Added test class for Artemis System and tested new method.

Updated display all method to account for instances in which the system
is unowned.