Implements the game loop and game setup sequence which allows players to run the game
setupGame is responsible for running the setup sequence, i.e. make sure the board is created, players have been created, and calling the randomisation method for the player order.

Updates createBoard method to throw an exception if no board object is passed through. The method will also return the initialised board object.

displayPurchasableComponent loop refactored to reduce duplicated code.

purchaseComponentOption updated to remove duplicated code. Some code commented out temporarily to be followed up as a TODO.

displayMenu method signature changed to require the current player object and a scanner object. A number of TODOs have been added to the method. At present none of the sub menu functionality is being invoked, in it's place comments are outputted to the players.

Removed the second definition of getNextPlayer as it is redundant.

playTurn has been removed, as gameLoop is serving in its place.

New overloaded version of announce method added. This will accept the currentPlayer object and will display outputs as "[currentPlayerName] message"