Implementation of Develop Components Menu and Bug Fixes
Added checkComponentCanBeDeveloped() to Compenent class. This method
returns a boolean true if the component in question can be developed.

Added a method in the Player class to return a Map of all their owned
components which can be developed.

Added getter to return developmentStagesMap. Useful for adding clarity
to the new menu showing components which can be developed by the player.

Added displayDevelopComponentMenu() and
displayComponentsPlayerCanDevelop() methods to facilitate players being
presented with components they can develop.

Linked methods checking if owner wishes to receive resources and
charging player for landing. The former will now call the latter.

Fixed a bug in which an exception was thrown when a player indicated
they wished to receive resources from the current player upon landing on
an owned square.

Fixed a bug in which a player would be asked if they wished to receive
resources from themselves upon landing on a square they themselves
owned.