Added check to update system owner and fixed logical error
Added functionality to setComponentOwner() to check if, following the
component owner taking charge of that particular component, they now own
all components in the system. If so, the method adds the relevant system
to the player's list of owned systems, and assigns the player as the
systemOwner of the system.

TODO - determine if this functionality is better placed elsewhere in the
system.

Logical fix - fixed an error in which players were only credited with
resources if the passed the recruitment square as opposed to landing on
it. Now, players should be credited with resources whether they land on
it or pass it.