Bug fixes, javadoc comment and implementing further checks
Updated javadoc comment for CompareByCounterOfResourcesDevoted to
specify it is used to order from high to low.

Added constant to determine longest possible player name length.
Adjusted createPlayers() to factor this in.

Adjusted gameLoop to incorporate checkIfSquareIsPurchasable(). Added
check within checkIfSquareIsPurchasable() to account for a player
landing on a square they themselves own.

Added announcements to provide better context when displaying components
player can develop. Added a check to output an announcement if player
lacks resources to perform a development.