Minor update to displayDevelopComponent
This update changes the logic in the displayDevelopComponent() method to
account for scenarios in which a player has the exact amount of
resources remaining to perform a development. Previously, they could
perform the development and be left with a balance of zero (an endgame
condition without triggering endgame). Now a message will be displayed
reminding the player that their resource balance must stay above zero or
the game will end.