Adds a new playerConfirmation method which replaces all method scanner loops asking for yes/no responses.
Code has been refactored to remove the duplication within multiple methods. All methods that require a yes or no response from the player are handled by getPlayerConfirmation(). This also helps keep consistency of messaging and inputs to and from the player.

getPlayerConfirmation() returns true if a player responds yes and false if no. Otherwise it will continue to loop until the player responds with a valid answer.

A private scanner variable has been created in the Game class, along with a getter and setter. This is directly accessible to all methods within Game, and can be accessed via the getter within the other classes.

Introducing this variable enables methods to access the scanner directly, without having to pass it through multiple methods. This allows for a cleaner code base and makes it easier to monitor.