Adds additional messaging to the players, simplifies logic, and adds unit tests
Removes redundant scanner - this is now set as a private class variable.

When a player does not have enough resources to purchase a component they will no longer be offered it. Instead they will be informed of this and game will automatically offer to other players.

Updated endGame wording to make it clear that Artemis was not launched (in the scenario of aborted mission).

Additional unit tests were added to ensure players with insufficient resources are not given the option to purchase a component.