Updates purchaseComponent and tradeComponent with associated unit testing
purchaseComponent now sets the owner of the system. This works by checking each purchase, when the final purchase of a component within a single system results in one owner having control of all systems, the systemOwner will be set. The system will be added to the player ownedSystems too.

tradeComponent performs the same logic but also includes removal of a system from ownedSystems, and setting systemOwner to null. This occurs when a systemOwner trades away a component, so they no longer own the system.

Additional unit tests were added to ensure that systemOwner is only set when ALL components within the system are owned by the same player. Checks also implemented so that when a component is traded away, that ownership reflects that no single person owns the system.