Adds new getters, javadocs, refactors method names and signatures, and reorders code for a more logical flow
New getters (isEndGame and isWinGame) added to enable unit testing for end of game sequences.

Javadocs added and updated throughout to provide more clarification as to the purpose of class variables and methods.

A small number of method signatures have been changed so that they return interfaces as opposed to hard implementations. This makes the code more flexible and easier to develop in future.

Method names have been refactored to make their purpose more obvious when reading through the code.

There are a small number of other code refactors to remove the possibility of game play exceptions.