Adds ability for the board object to manage developedSystems with associated unit test.
Board now has a private list for storing systems that have been developed (i.e. all components within the system are fully developed).

Systems are automatically detected as developed through Component.setDevelopmentStage which automatically invokes Board.addDevelopedSystem. Which then checks if the system should is developed.

If the system is developed, it is added to developedSystems list and the method returns true. Otherwise it returns false, and developedSystems is unaltered.

Unit test checks that system is only considered developed when all components that reside within it are fully developed.