Refactors how developmentStage is updated
Updated implementation means that developmentStage should only ever be updated using setDevelopmentStage. This is so that Board.addDevelopedSystem is invoked to check if the system is fully developed.

developComponent has been refactored to update the developStage at the very end of the method. This is so that the order of announcements is correct, i.e. the component announcement takes place before the system announcement (if relevant).

incrementDevelopmentStage no longer updates the instance variable value, instead it returns what the updated stage should be. This is captured in a variable within developComponent and set at the end of the method.