Updates playTurn exception handling and adds setter for gameWon variable
The exception handling for playTurn was resulting in an infinite loop when unexpected values were entered. The updated coded resolves this.

A setter was needed for the variable gameWon in order to reset the value after each test was run. Some tests were modified the value, which persisted and caused other tests to fail. The value is now set before each test.